+++ agent.h
@@ -2,12 +2,12 @@
 
 typedef struct _node {
    Category *tree;
-   float h, t;
+   double h, t;
    int name_mod, time_stamp;
 } Agent;
 
 Agent *create_agent(void);
-bool negotiate(Agent *spk, Agent *lst, float x, float y, int time, int *split_count);
-float overlap(Agent *agent_x, Agent *agent_y, float min, float max, bool env);
+bool negotiate(Agent *spk, Agent *lst, double x, double y, int time, int *split_count);
+double overlap(Agent *agent_x, Agent *agent_y, double min, double max, bool env);
 Agent *clone_agent(Agent *agent);
 void delete_agent(Agent *agent);
